Skip using issues as layouts
